+++ pkgs/google_cloud_storage/test/get_object_acl_test.dart
@@ -20,7 +20,7 @@ import 'package:test_utils/cloud.dart' as cloud;
 
 import 'test_utils.dart';
 
-void deleteObjectAclTest(Storage Function() createStorage) {
+void getObjectAclTest(Storage Function() createStorage) {
   late Storage storage;
 
   setUp(() {
@@ -32,7 +32,7 @@ void deleteObjectAclTest(Storage Function() createStorage) {
   test('success', () async {
     final bucketName = await createBucketWithTearDown(
       storage,
-      'del_obj_acl_ok',
+      'get_obj_acl_ok',
       metadata: BucketMetadata(
         iamConfiguration: BucketIamConfiguration(
           uniformBucketLevelAccess: UniformBucketLevelAccess(enabled: false),
@@ -50,29 +50,19 @@ void deleteObjectAclTest(Storage Function() createStorage) {
     const entity = 'user-${cloud.googleTestUser}';
     await storage.insertObjectAcl(bucketName, 'object.txt', entity, 'READER');
 
-    await storage.deleteObjectAcl(bucketName, 'object.txt', entity);
-
-    final metadata = await storage.objectMetadata(
-      bucketName,
-      'object.txt',
-      projection: 'full',
-    );
-
-    final testUserRoles = [
-      for (var i in metadata.acl ?? <ObjectAccessControl>[])
-        if (i.entity == entity) (i.entity, i.role),
-    ];
-    expect(testUserRoles, isEmpty);
+    final acl = await storage.getObjectAcl(bucketName, 'object.txt', entity);
+    expect(acl.entity, entity);
+    expect(acl.role, 'READER');
   });
 
   test('no object', () async {
     final bucketName = await createBucketWithTearDown(
       storage,
-      'del_obj_acl_no_object',
+      'get_obj_acl_no_object',
     );
 
     expect(
-      () => storage.deleteObjectAcl(bucketName, 'non-existent.txt', 'allUsers'),
+      () => storage.getObjectAcl(bucketName, 'non-existent.txt', 'allUsers'),
       throwsA(isA<NotFoundException>()),
     );
   });
@@ -80,7 +70,7 @@ void deleteObjectAclTest(Storage Function() createStorage) {
   test('no acl', () async {
     final bucketName = await createBucketWithTearDown(
       storage,
-      'del_obj_acl_no_acl',
+      'get_obj_acl_no_acl',
       metadata: BucketMetadata(
         iamConfiguration: BucketIamConfiguration(
           uniformBucketLevelAccess: UniformBucketLevelAccess(enabled: false),
@@ -97,7 +87,7 @@ void deleteObjectAclTest(Storage Function() createStorage) {
 
     const entity = 'user-${cloud.googleTestUser}';
     expect(
-      () => storage.deleteObjectAcl(bucketName, 'object.txt', entity),
+      () => storage.getObjectAcl(bucketName, 'object.txt', entity),
       throwsA(isA<NotFoundException>()),
     );
   });
@@ -105,7 +95,7 @@ void deleteObjectAclTest(Storage Function() createStorage) {
   test('success with generation', () async {
     final bucketName = await createBucketWithTearDown(
       storage,
-      'del_obj_acl_gen_ok',
+      'get_obj_acl_gen_ok',
       metadata: BucketMetadata(
         iamConfiguration: BucketIamConfiguration(
           uniformBucketLevelAccess: UniformBucketLevelAccess(enabled: false),
@@ -123,30 +113,20 @@ void deleteObjectAclTest(Storage Function() createStorage) {
     const entity = 'user-${cloud.googleTestUser}';
     await storage.insertObjectAcl(bucketName, 'object.txt', entity, 'READER');
 
-    await storage.deleteObjectAcl(
+    final acl = await storage.getObjectAcl(
       bucketName,
       'object.txt',
       entity,
       generation: md.generation,
     );
-
-    final metadata = await storage.objectMetadata(
-      bucketName,
-      'object.txt',
-      projection: 'full',
-    );
-
-    final testUserRoles = [
-      for (var i in metadata.acl ?? <ObjectAccessControl>[])
-        if (i.entity == entity) (i.entity, i.role),
-    ];
-    expect(testUserRoles, isEmpty);
+    expect(acl.entity, entity);
+    expect(acl.role, 'READER');
   });
 
   test('not found with wrong generation', () async {
     final bucketName = await createBucketWithTearDown(
       storage,
-      'del_obj_acl_gen_not_found',
+      'get_obj_acl_gen_not_found',
       metadata: BucketMetadata(
         iamConfiguration: BucketIamConfiguration(
           uniformBucketLevelAccess: UniformBucketLevelAccess(enabled: false),
@@ -165,7 +145,7 @@ void deleteObjectAclTest(Storage Function() createStorage) {
     await storage.insertObjectAcl(bucketName, 'object.txt', entity, 'READER');
 
     expect(
-      () => storage.deleteObjectAcl(
+      () => storage.getObjectAcl(
         bucketName,
         'object.txt',
         entity,
@@ -177,9 +157,9 @@ void deleteObjectAclTest(Storage Function() createStorage) {
 }
 
 void main() async {
-  group('delete object acl', () {
+  group('get object acl', () {
     group('google-cloud', tags: ['google-cloud', 'no-ulba'], () {
-      deleteObjectAclTest(Storage.new);
+      getObjectAclTest(Storage.new);
     });
 
     group('storage-testbench', tags: ['storage-testbench'], () {
@@ -191,27 +171,7 @@ void main() async {
 
       tearDown(() => storage.close());
 
-      deleteObjectAclTest(() => storage);
-    });
-
-    test('non-idempotent transport failure', () async {
-      var count = 0;
-      final mockClient = MockClient((request) async {
-        count++;
-        if (count == 1) {
-          throw http.ClientException('Some transport failure');
-        } else {
-          throw StateError('Unexpected call count: $count');
-        }
-      });
-
-      final storage = Storage(client: mockClient, projectId: 'fake project');
-
-      await expectLater(
-        storage.deleteObjectAcl('bucket', 'object', 'entity'),
-        throwsA(isA<http.ClientException>()),
-      );
-      expect(count, 1);
+      getObjectAclTest(() => storage);
     });
 
     test('idempotent transport failure', () async {
@@ -221,7 +181,11 @@ void main() async {
         if (count == 1) {
           throw http.ClientException('Some transport failure');
         } else if (count == 2) {
-          return http.Response('', 204);
+          return http.Response(
+            '{"kind": "storage#objectAccessControl", '
+            '"entity": "user-test", "role": "READER"}',
+            200,
+          );
         } else {
           throw StateError('Unexpected call count: $count');
         }
@@ -229,12 +193,9 @@ void main() async {
 
       final storage = Storage(client: mockClient, projectId: 'fake project');
 
-      await storage.deleteObjectAcl(
-        'bucket',
-        'object',
-        'entity',
-        generation: BigInt.from(1),
-      );
+      final acl = await storage.getObjectAcl('bucket', 'object', 'user-test');
+      expect(acl.entity, 'user-test');
+      expect(acl.role, 'READER');
       expect(count, 2);
     });
   });